+++ database/migrations/2022_01_27_103111_create_commissions_table.php
@@ -16,7 +16,7 @@ class CreateCommissionsTable extends Migration
         Schema::create('commissions', function (Blueprint $table) {
             $table->id();
 
-            $table->double('amount',6,2);
+            $table->decimal('amount',6);
             $table->string('type'); //SELF / REFERRAL / SYSTEM
             $table->foreignId('account_id')->unsigned()->nullable();
             $table->foreignId('transaction_id')->constrained()->cascadeOnDelete();
