+++ database/migrations/2022_01_27_111750_create_float_account_transactions_table.php
@@ -17,7 +17,7 @@ class CreateFloatAccountTransactionsTable extends Migration
             $table->id();
 
             $table->string('type', 20); // DEBIT or CREDIT
-            $table->double('amount');
+            $table->decimal('amount');
             $table->string('description');
             $table->foreignId('float_account_id')->constrained();
